Initialize as mix project
So this can be used as hex package.

Still need to fix the ConnHelpers module to use Floki's functions
instead of the imported functions from an out-of-scope module; also need
to use the Endpoint of the project that this package will be used in;
ans need to create tests for the ConnHelpers module.