Add possibility to use the module attribute @endpoint from test case
Introduce a `__using__/1` macro to import TestDispatchForm and two macro
functions: `dispatch_form_with/3` and `dispatch_form_with/2` to pass the
@endpoint module attribute, that should have been set in the test case,
over to the non-macro function `dispatch_form_with/4`. This way we can
make use of the @endpoint module attribute that is also used by
`Phoenix.ConnTest` and `Phoenix.ChannelTest`, and there is no need to
set an Application configuration.

Also adjust `send_to_action` to use this other endpoint, and add a check
in `dispatch_form_with/4` if the @endpoint is set, and raise with a
meaningful message if not.

Comment out the @spec for `dispatch_form_with/4` because this should
still be adjusted to comply with the new arity of the function.

Should also adjust the documentation/readme on this part.