Add possibility to use the module attribute `:endpoint` from test case
Introduce the function `retrieve_endpoint/1` to either get the
application config `:endpoint` or the conn's endpoint to be used for
`Phoenix.ConnTest.dispatch/5`. This way it's possible to use the
`:endpoint` module attribute (also used by `Phoenix.ConnTest` and
`Phoenix.ChannelTest`) without adding an application config setting.

Also adjust the documentation and README.md to explain this.