dispatch links
Some actions on a page might be in forms of <a href=""></a> tags. With
this helper they can be found and call the links. With this feature the actual
links can be called from controller tests, and therefore they are
testing whether the links are doing what they should do.

This commit introduces TestDispatch module where the public functions
live (at this moment dispatch_form_with and dispatch_link.
The two other modules TestDispatch.Form and TestDispatch.Link are helper
modules.

A couple of routes and pages are added to test the flow of dispatch_link