dipatch_form/4 can use both entity and test_selector
When a form is slected with a test_selector but the items are using the
entity for each field we want to split up the attributes in a map within
a map.

Introduces dispatch_form/4 that with the entity as third and
test_selector as fourth param. It will find the right form by the
test_selector and uses the entity to find the inputs and update the
right values