Match nestings when an entity is not given
When submitting a form with the structure `course[subject][part]`
is set as the name of a field a typicle match in the controller could be
```ex
def post(conn, %{"course" => %{"subject" => %{"part" => part_value}}}) do
```

But this is not send by us, instead what we sent will resolve as
`%{"course_subject_part" => part_value}`.

In this commit I'm trying to solve this by getting the keys from the
name attribute instead of the id. This is easier to match on and the
regex can be replaced with a simple `String.split/3`

I was a bit confused by the tests that i've adjusted, where we match the
params with `cats => %{ ... }`. Before this commit there weren't any
numbers even though they were specified and to my understanding they
should be there as they are now. Please let me know if I missed
something here.